MODCLUSTER-754 Modify ResetRequestSourceImpl to send STOP request when proxy not aware of a context
When an app server connects to a proxy for the first time or after a
disruption, it sends an INFO request to the proxy. Proxy responds with a
list of contexts that are registered for given server, and their states.
The context states returned from the proxy are compared to the actual
states of contexts on the app server, and if necessary, the server
should send ENABLE / STOP requests to bring the proxy to an up-to-date
state.

The change in this commit makes the app server send a STOP request for a
context in case when the proxy response to an INFO request doesn't
contain given context. The purpose of this STOP request is to enforce
context registration on the proxy side.